postas enla pista

+++ tp/common/Posta.h
@@ -16,7 +16,7 @@ public:
   virtual void Contact(Contactable* contactable);
   virtual void GetContactedBy(Car* car);
   virtual void GetContactedBy(Posta* posta);
-  void Place(b2World& world, b2Vec2 position, int angle);
+  void Place(b2World& world, b2Vec2 position, float32 angle);
   int GetId();
   b2Vec2 GetPosition();
   int GetAngle();
